net: lib: azure_iot_hub: Send `_FOTA_DONE` event if PUBACK is received.
Notify the application that it is safe to perform a reboot with
`AZURE_IOT_HUB_EVT_FOTA_DONE` after receiving a PUBACK that the
last FOTA report was sent.

If the application shut down too fast there is a risk that the last
FOTA report sent after successfully downloading a new firmware
image is not sent.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>