net: nrf_cloud_pgps: Add payload to P-GPS events
This patch reworks P-GPS events to contain a structure with the
event type and optional payload.

The change is needed for the new PGPS_EVT_REQUEST that is also
introduced.
This new event type will be received if CONFIG_NRF_CLOUD_MQTT is
disabled and one of the following public APIs is called:
- nrf_cloud_pgps_preemptive_updates()
- nrf_cloud_pgps_request_all()
- nrf_cloud_pgps_init() and the device has an incomplete set of
  predictions stored.

The event has payload of type struct nrf_cloud_pgps_request,
and the intention is that the application can use this information
and the transport of its choice to request P-GPS data from nRF Cloud.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>