applications: Asset Tracker_v2: Split `prj.conf` into multiple overlays
Split `prj.conf` into multiple overlay configuration files where
each file contains options that are specific to a feature.

To build with a specific feature the corresponding overlay file
needs to be included in the `OVERLAY_CONFIG` CMake list-variable when
building the application.

This patch is to provide a more modular approach compared to having all
configurations in the root `prj.conf` and relying on the user to
comment out/in options when wanting to enable a feature.

CIA-390 CIA-407

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>