net: lib: nrf_cloud: Add Kconfig option to fetch missing P-GPS data
Add Kconfig option that decides wether the nRF Cloud P-GPS library
should fetch missing predictions upon initialization of the library.

Enabling this option require that the library is initialized when
a connection to cloud has been established beforehand.

This is not the case in the Asset Tracker v2, hence the option is
by default disabled when enabling P-GPS via the
`overlay-pgps.conf` overlay.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>