SQUASH: wifi_provision: Add APIs to grant/deny access to internal resources
Add APIs to grant/deny access to internal resources:
- wifi_provision_client_access_grant
- wifi_provision_client_access_deny

When a STA connects the library will now persist its MAC and
only grant access to its internal HTTP provisioning resources if
granted by the library caller.

The user needs to know when its provisioning client is
connected to the softAP when it grants access to avoid
potential malicious clients gaining access to the internal resources.

This functionality depends on a requested feature to limit the softAP
stack to only support a single connected STA at the time.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>